requirements.txt
GPXpert~=0.0.post1.dev1+g5cab12b
pytest~=8.2.2
setuptools~=70.1.0
pytest-cov~=5.0.0
pytest-sugar~=1.0.0
gpxpy~=1.6.2
srtm.py~=0.3.7